Implement displaying the white card selection embed

+++ src/main/kotlin/dev/vrba/discord/cah/entities/Game.kt
@@ -25,7 +25,7 @@ data class Game(
     val blackCard: Int? = null,
 
     @Column("judge_id")
-    val judge: Int? = null,
+    val judge: Long? = null,
 
     @Column("used_white_cards")
     val usedWhiteCards: List<Int> = emptyList(),
